feat(EventDetails): Add user interactions
The Unsubscribe button triggers unsubscribeFromEvent request.
Shown to subscribers when event is notYetScheduled.

The Delete button triggers removeEvent request.
Shown to organizer when event is either pending or notYetScheduled.

The Subscriptions component have an IconButton next to each subscriber
which triggers removeUserFromEvent request.
Shown to organizer when event is either pending or notYetScheduled.